Removed regexp, time, strings, cycles usage. It's very time-consuming. 

before:
```
goos: darwin
goarch: amd64
pkg: github.com/ngalayko/go
BenchmarkValid-4         1000000              1322 ns/op
PASS
ok      github.com/ngalayko/go  1.355s
```

after:
```
goos: darwin
goarch: amd64
pkg: github.com/ngalayko/go
BenchmarkValid-4                20000000               100 ns/op
BenchmarkValidString-4          20000000                82.7 ns/op
PASS
ok      github.com/ngalayko/go  3.877s
```