@frozzare 2018-11-29
Thanks, looks good! Haven't updated this package so much since 2014 so the code is quite old :)