@ngalayko 2018-11-29
yea, you are right! I will update it with tests

> On 29 Nov 2018, at 20:33, Calle Pettersson <notifications@github.com> wrote:
> 
> @carlpett commented on this pull request.
> 
> In personnummer.go <https://github.com/personnummer/go/pull/5#discussion_r237628758>:
> 
> >  	}
>  
> -	return day
> +	if year%4 == 0 {
> This actually doesn't cover the full leap year logic - apart from being divisible by four, it also needs to not be divisible by 100 unless it is also divisible by 400.
> For example, 1900 was not a leap year, but 2000 was.
> 
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub <https://github.com/personnummer/go/pull/5#pullrequestreview-179975774>, or mute the thread <https://github.com/notifications/unsubscribe-auth/AOUJ9BghaEh-LRi1TfhOKCJxO-ppJsaPks5u0DaUgaJpZM4Y51X9>.
>